apps/directory-pipeline/requirements.txt
# Attorney Directory Pipeline - California
psycopg2-binary>=2.9.9
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
